Move configuration snapshot based periodic rules into old-periodic/ directory. Update README and RULES to indicate as such.
