+++ old-periodic/iam_password_require_symbol-periodic.js
